successfully save all Config information in dataBase.

+++ common/src/main/java/com/nima/common/database/entitty/URLIdSecond.kt
@@ -8,6 +8,6 @@ data class URLIdSecond(
     @PrimaryKey var id: Long? = 0,
     var urlId: Long? = null,
     var regex : String? = null,
-    var startIndex : Int? = null,
-    var finishIndex : Int? = null,
+    var startIndex : Long? = null,
+    var finishIndex : Long? = null,
 )
